audit(#1237): address coderabbit review comments — checked casts + long metadata
addresses 60+ unresolved review threads on pr #1244 with three changes:

1. **`modelmetadata.complexity` widened to long**. complexity reports
   the model's parameter count for ml dashboards; for foundation-scale
   diffusion models (sora, hidream full, sd3.5 large, hunyuanvideo,
   flux 2) the int field would silently truncate. featurecount stays
   int (it's input-feature count, not parameter count, and shouldn't
   change semantics — wrapped via `(int)math.min((long)int.maxvalue, ...)`
   at the assignment sites that use parametercount for it).

2. **bulk replace `int X = (int)Y.parametercount` with
   `checked((int)Y.parametercount)`** in setparameters slicing paths.
   reviewers flagged that silent overflow on foundation-scale wrapped
   sub-models (predictor, vae, controlnet encoder, lora adapters)
   would corrupt slicing behavior — checked() throws explicitly so
   the caller sees the invalid model state rather than a silent
   buffer-size mismatch downstream. covers ~84 files in fastgeneration,
   imageediting, control, video, virtualtryon, audio, etc.

3. **federated dto cast removed**. `federatedrunparametersresponse.parametercount`
   and `federatedrunstatusresponse.parametercount` are already long;
   the `(int)state.parametercount` cast at the assignment site was a
   leftover that defeated the dto's widening. dropped both casts.

4. **complexity = (int)parametercount → complexity = parametercount**.
   no narrowing needed since complexity is now long. metadata sites
   that used the (int) cast pre-widening now pass through cleanly.

5. **redundant double-casts `(int)((int)x.parametercount)`** dropped
   (artifact from earlier sed pass).

6. **uvit/controlnet union pro int accumulators widened to long**.
   same foundation-scale silent-overflow concern as ditnoisepredictor /
   mmditnoisepredictor; `int count = 0` → `long count = 0` and drop
   per-layer (int) casts on the running sum.

7. **inconsistent `getparameters().length` replaced with
   `parametercount`** in jen1 and t2iadapter setparameters — getparameters
   materialises the entire flat aggregate just to read its length
   (wasteful at foundation scale).

build status: src + tests clean (0 errors).

+++ src/AiDotNet.Serving/Services/Federated/FederatedCoordinatorService.cs
@@ -195,7 +195,7 @@ public sealed class FederatedCoordinatorService : IFederatedCoordinatorService
             {
                 RunId = state.RunId,
                 RoundNumber = state.CurrentRound,
-                ParameterCount = (int)state.ParameterCount,
+                ParameterCount = state.ParameterCount,
                 Parameters = state.GlobalParameters.ToArray()
             };
         }
@@ -300,7 +300,7 @@ public sealed class FederatedCoordinatorService : IFederatedCoordinatorService
             {
                 RunId = state.RunId,
                 CurrentRound = state.CurrentRound,
-                ParameterCount = (int)state.ParameterCount,
+                ParameterCount = state.ParameterCount,
                 JoinedClients = state.JoinedClients.Count,
                 UpdatesReceivedForCurrentRound = state.PendingClientParameters.Count,
                 MinUpdatesRequired = state.MinClientUpdatesPerRound
