+++ src/ReinforcementLearning/Agents/EveryVisitMonteCarloAgent.cs
@@ -241,7 +241,7 @@ public class EveryVisitMonteCarloAgent<T> : ReinforcementLearningAgentBase<T>
 
     public override ModelMetadata<T> GetModelMetadata()
     {
-        return new ModelMetadata<T> { FeatureCount = this.FeatureCount, Complexity = (int)ParameterCount };
+        return new ModelMetadata<T> { FeatureCount = this.FeatureCount, Complexity = ParameterCount };
     }
 
     public override long ParameterCount =>
